adapt the minigpt jaxaistack example for text generation
The key trick for a single batch jit-able generation is to
slice logits[0][sample_index] instead of
logits[:, sample_index, :] to avoid dynamic shapes.

The downside is that this treats one batch at a time but
for toy text generation this is acceptable.

Plus you get the animation printing using it this way
compared to my current generate_fast method
implementation.

Ref: https://docs.jaxstack.ai/en/latest/JAX_for_LLM_pretraining.html