fix newline token prediction issue
Line 410 in model.py is the culprit!

I was computing logits for the entire
answer-length-padded sequence which
is way bigger then block_size. This means
the logits has NaNs for block dimension
> block_size.

I THOUGHT IT WAS A MODEL COLLAPSE
ISSUE BUT IT WASN'T!

Lesson learnt: Always check for NaNs in logits.